time type int to bigint because miliseconds required

+++ backend/src/database/models/gps.py
@@ -1,4 +1,4 @@
-from sqlalchemy import ForeignKey, Integer, String
+from sqlalchemy import BigInteger, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from src.database import DB_SCHEMA, Base
@@ -10,7 +10,7 @@ class GPS(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     device_id: Mapped[int] = mapped_column(Integer, ForeignKey(f"{DB_SCHEMA}.device.id"), nullable=False)
-    time: Mapped[int] = mapped_column(Integer, nullable=True)
+    time: Mapped[int] = mapped_column(BigInteger, nullable=True)
     sentence: Mapped[str] = mapped_column(String(255), nullable=False)
 
     device: Mapped["Device"] = relationship(back_populates="gps_readings")
